added mz 2 frequency modeling
With the discovery that frequency = int + slope / mz^0.5, we need the tooling to enable that use, and get rid of the local linear fitting.

So, we have functions for creating the fit and applying it, and then we are using them, and storing the coefficients from the model for subsequent use if we need them.

This is so much faster than before! Every step in here is vectorized (for a given scan), so it just goes lickety split!