the offset used to be derived from the model of mz differences to mz, but we don't officially use that anymore, but it was useful for SMIRFE assignment
So now we determine the good points again, get the median difference in frequency space from previously calculated frequency values, and then build a model of mz ~ frequency, and get the difference in mz space from the difference in frequency space. A graph of the offset vs mz shows the expected behavior.

Also sorting in ascending mz to make debugging SMIRFE assignment easier